refactor: resetPrefCodesに名前を変更したので、パスを修正した
